+++ src/pages/index.js
@@ -5,8 +5,8 @@ import {
   addCardBtn,  
   apiConfig,
   validationConfig,
-} from "../components/variables.js";
-import Api from "../components/api.js";
+} from "../utils/variables.js";
+import Api from "../components/Api.js";
 import Card from "../components/Card.js";
 import FormValidator from "../components/FormValidator.js";
 import PopupWithForm from "../components/PopupWithForm.js";
